dockers/Validador/requirements.txt
attrs==23.1.0
bcrypt==3.1.7
blinker==1.6.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==2.8
dnspython==2.3.0
entrypoints==0.3
fasteners==0.14.1
Fiona==1.9.3
Flask==2.3.2
future==0.18.2
geopandas==0.13.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keyring==18.0.1
Mako==1.1.0
MarkupSafe==2.1.2
monotonic==1.5
munch==2.5.0
netifaces==0.10.4
numba==0.57.0
numpy==1.24.3
oauthlib==3.1.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.6.0
patsy==0.5.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.6
pymongo==4.3.3
python-dateutil==2.8.2
pyts==0.12.0
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==2.3.1
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
statsmodels==0.14.0
threadpoolctl==3.1.0
urllib3==1.25.8
Werkzeug==2.3.4
zipp==3.15.0